requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-appconf==1.0.2
django-cors-headers==2.1.0
django-extensions==1.8.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
jsonfield==2.0.2
MarkupSafe==1.0
numpy==1.13.1
PyJWT==1.5.2
python-slugify==1.1.4
pytz==2017.2
six==1.10.0
Unidecode==0.4.21
